多处更新
1. 修正timer接口，利用C++11 chrono，更加符合规范
2. 增加socket pool，降低socket创建、销毁开销，增强系统性能
3. cancel接口在win上实现喂cancelioex，不支持XP

+++ include/async_io/timer/impl/timer_impl.hpp
@@ -16,7 +16,7 @@ namespace async { namespace timer {
 	private:
 		HANDLE timer_;
 		long period_;		// 间隔时间
-		long due_;			// 第一次后延长时间
+		long long due_;		// 延长时间
 
 	public:
 		waitable_timer_t(long period, long due, bool manualReset = false, const wchar_t *timerName = nullptr);
@@ -47,7 +47,7 @@ namespace async { namespace timer {
 		}
 
 	public:
-		void set_timer(long period, long delay);
+		void set_timer(long period, long long delay);
 
 		void cancel();
